requirements.txt

ipython
matplotlib
numpy
opencv-python
pandas
pillow
psutil
scikit-learn
scipy
seaborn
torch
torchvision
#transformers
tqdm
ultralytics

